requirements.txt
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
protobuf>=3.20.0
six>=1.16.0
tensorboardX>=2.5
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
opencv-python>=4.8.0
imageio>=2.31.0
